WebStudio/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
environs==9.5.0
psycopg2==2.9.3
loguru==0.6.0
Pillow==9.1.0